+++ testsAT/src/test/java/com/stratio/cassandra/lucene/testsAT/util/CassandraConfig.java
@@ -24,9 +24,9 @@ public class CassandraConfig {
 
     static final boolean EMBEDDED = Boolean.parseBoolean(get("embedded", "true"));
     static final String HOST = get("host", "127.0.0.1");
-    static final int REPLICATION = Integer.valueOf(get("replication", "1"));
+    static final int REPLICATION = Integer.valueOf(get("replication", "3"));
     static final ConsistencyLevel CONSISTENCY = ConsistencyLevel.valueOf(get("consistency", "QUORUM"));
-    static final int FETCH = Integer.parseInt(get("fetch", "100"));
+    static final int FETCH = Integer.parseInt(get("fetch", "10"));
     static final int THREADS = Integer.parseInt(get("threads", "0"));
     static final int REFRESH = Integer.parseInt(get("refresh", "1"));
     static final int WAIT_FOR_INDEXING = Integer.parseInt(get("wait_for_indexing", "1"));